backend/requirements.txt
annotated-types
anyio
beautifulsoup4
blis
cachetools
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
confection
cryptography
cymem
deprecation
Django
django-cors-headers
djangorestframework
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
h11
h2
hpack
httpcore
httplib2
httpx
hyperframe
idna
Jinja2
joblib
langcodes
language-data
lxml
marisa-trie
markdown-it-py
MarkupSafe
mdurl
murmurhash
nltk
numpy
packaging
pandas
pdfplumber
pillow
postgrest
preshed
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
Pygments
PyJWT
pyparsing
PyPDF2
pypdfium2
python-dateutil
python-docx
python-dotenv
pytz
realtime
regex
requests
rich
rsa
scikit-learn
scipy
setuptools
shellingham
six
smart-open
soupsieve
spacy
spacy-legacy
spacy-loggers
sqlparse
srsly
storage3
StrEnum
supabase
supabase-auth
supabase-functions
thinc
threadpoolctl
tqdm
typer
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
wasabi
weasel
websockets
wrapt
grpcio==1.71.0
grpcio-status==1.71.0
pdfminer.six==20240706
gunicorn
whitenoise

